:recycle: refactor: 클릭수 봇 여부 판별 로직 분리
